+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+☮️ Peace in the world, or the world in pieces. 🕊️
 
 ## [2.5.0] - 2024-03-24
-☮️ Peace in the world, or the world in pieces. 🕊️
 ### Added
 - New SWIM config for actors allowing the user to set up a custom mirror self token image. If using the Mighty Summoner, this image will be used for the mirrored self. If empty it will use the summoners token instead. This setting *does* accept wild card images.
 - Added shaken link to official module link list.